Creacion de clases controladoras y otros ajustes para iniciar la creacion de la interfaz grafica

+++ ProyectoFinalCodigo/src/main/java/org/uniquindio/proyectofinalcodigo/model/Visitante.java
@@ -189,7 +189,7 @@ public class Visitante extends Persona {
 
     // Pagar usando Saldo Virtual
 
-    public boolean pagarConSaldoVirtual(int cantidad){
+    public boolean pagarConSaldoVirtual(double cantidad){
 	    if(visitanteactivo.getSaldoVirtual()>cantidad){
 		    visitanteactivo.setSaldoVirtual(visitanteactivo.getSaldoVirtual()-cantidad);
 		    return true;
